+++ back-end/src/main/java/com/arkon/exam/model/MetrobusUnitLocation.java
@@ -1,4 +1,4 @@
-package com.arkon.exam.web;
+package com.arkon.exam.model;
 
 import java.sql.Timestamp;
