+++ back-end/src/main/java/com/arkon/exam/repository/MetrobusUnitLocationRepository.java
@@ -1,6 +1,8 @@
-package com.arkon.exam.web;
+package com.arkon.exam.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+
+import com.arkon.exam.model.MetrobusUnitLocation;
 
 public interface MetrobusUnitLocationRepository extends JpaRepository<MetrobusUnitLocation, Integer>{
