+++ back-end/src/main/java/com/arkon/exam/repository/TownRepository.java
@@ -1,6 +1,8 @@
-package com.arkon.exam.web;
+package com.arkon.exam.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+
+import com.arkon.exam.model.Town;
 
 public interface TownRepository extends JpaRepository<Town, Integer>{
